# docs/swagger.yaml
...
info:
  description: The Venice.ai API.
  termsOfService: https://venice.ai/legal/tos
  title: Venice.ai API
  version: '20250713.224148'
openapi: 3.0.0
security:
- BearerAuth: []
servers:
- url: https://api.venice.ai/api/v1
...
components:
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
  schemas:
    StandardError:
      type: object
      properties:
        error:
          type: string
          description: A description of the error
      required:
      - error
    DetailedError:
      type: object
      properties:
        details:
          type: object
          properties: {}
          description: Details about the incorrect input
          example:
            _errors: []
            field:
              _errors:
              - Field is required
        error:
          type: string
          description: A description of the error
      required:
      - error
...
paths:
...
  /api_keys:
    get:
      description: Return a list of API keys.
      operationId: getApiKeys
      summary: /api/v1/api_keys
      tags:
      - API Keys
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        apiKeyType:
                          type: string
                          enum:
                          - INFERENCE
                          - ADMIN
                          description: API Key type
                          example: ADMIN
                        consumptionLimits:
                          type: object
                          properties:
                            usd:
                              anyOf:
                              - type: number
                                minimum: 0
                              - nullable: true
                                title: 'null'
                              - nullable: true
                                title: 'null'
                              description: USD limit
                              example: 50
                            diem:
                              anyOf:
                              - type: number
                                minimum: 0
                              - nullable: true
                                title: 'null'
                              - nullable: true
                                title: 'null'
                              description: Diem limit
                              example: 10
                            vcu:
                              anyOf:
                              - type: number
                                minimum: 0
                              - nullable: true
                                title: 'null'
                              - nullable: true
                                title: 'null'
                              description: VCU limit (deprecated - use Diem instead)
                              deprecated: true
                              example: 100
                          description: The API Key consumption limits for each epoch.
                          example:
                            usd: 50
                            diem: 10
                        createdAt:
                          type: string
                          nullable: true
                          description: API Key creation date
                          example: 2023-10-01 12:00:00+00:00
                        description:
                          type: string
                          description: API Key description
                          example: Example API Key
                        expiresAt:
                          type: string
                          nullable: true
                          description: API Key expiration date
                          example: 2023-10-01 12:00:00+00:00
                        id:
                          type: string
                          description: API Key ID
                          example: e28e82dc-9df2-4b47-b726-d0a222ef2ab5
                        last6Chars:
                          type: string
                          description: Last 6 characters of the API Key
                          example: 2V2jNW
                        lastUsedAt:
                          type: string
                          nullable: true
                          description: API Key last used date
                          example: 2023-10-01 12:00:00+00:00
                        usage:
                          type: object
                          properties:
                            trailingSevenDays:
                              type: object
                              properties:
                                usd:
                                  type: string
                                  description: USD usage in the trailing 7 days
                                  example: '10.2424'
                                vcu:
                                  type: string
                                  description: VCU usage in the trailing 7 days (deprecated
                                    - use Diem instead)
                                  deprecated: true
                                  example: '42.2315'
                                diem:
                                  type: string
                                  description: Diem usage in the trailing 7 days
                                  example: '4.2231'
                              required:
                              - usd
                              - vcu
                              - diem
                          required:
                          - trailingSevenDays
                      required:
                      - apiKeyType
                      - consumptionLimits
                      - createdAt
                      - expiresAt
                      - id
                      - last6Chars
                      - lastUsedAt
                      additionalProperties: false
                    description: List of active API keys
                  object:
                    type: string
                    enum:
                    - list
                required:
                - data
                - object
                additionalProperties: false
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: An unknown error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
    delete:
      description: Delete an API key.
      operationId: deleteApiKey
      parameters:
      - description: The ID of the API key to delete
        in: query
        name: id
        required: false
        schema:
          type: string
      summary: /api/v1/api_keys
      tags:
      - API Keys
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                - success
                additionalProperties: false
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedError'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: An unknown error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
    post:
      description: Create a new API key.
      operationId: createApiKey
      summary: /api/v1/api_keys
      tags:
      - API Keys
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                apiKeyType:
                  type: string
                  enum:
                  - INFERENCE
                  - ADMIN
                  description: The API Key type. Admin keys have full access to the
                    API while inference keys are only able to call inference endpoints.
                  example: ADMIN
                consumptionLimit:
                  type: object
                  properties:
                    usd:
                      anyOf:
                      - type: number
                        minimum: 0
                      - nullable: true
                        title: 'null'
                      - nullable: true
                        title: 'null'
                      description: USD limit
                      example: 50
                    diem:
                      anyOf:
                      - type: number
                        minimum: 0
                      - nullable: true
                        title: 'null'
                      - nullable: true
                        title: 'null'
                      description: Diem limit
                      example: 10
                    vcu:
                      anyOf:
                      - type: number
                        minimum: 0
                      - nullable: true
                        title: 'null'
                      - nullable: true
                        title: 'null'
                      description: VCU limit (deprecated - use Diem instead)
                      deprecated: true
                      example: 100
                  description: The API Key consumption limits for each epoch.
                  example:
                    usd: 50
                    diem: 10
                description:
                  type: string
                  description: The API Key description
                  example: Example API Key
                expiresAt:
                  anyOf:
                  - type: string
                    enum:
                    - ''
                  - type: string
                    pattern: ^\d{4}-\d{2}-\d{2}$
                  - type: string
                    pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d{3})?Z$
                  description: The API Key expiration date. If not provided, the key
                    will not expire.
                  example: 2023-10-01 12:00:00+00:00
              required:
              - apiKeyType
              - description
              additionalProperties: false
              description: The request body for creating a new API key. API key creation
                is rate limited to 20 requests per minute and a maximum of 500 active
                API keys per user. VCU is being deprecated in favor of Diem. Please
                update your API calls to use Diem instead.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      apiKey:
                        type: string
                        description: The API Key. This is only shown once, so make
                          sure to save it somewhere safe.
                      apiKeyType:
                        type: string
                        enum:
                        - INFERENCE
                        - ADMIN
                        description: The API Key type
                        example: ADMIN
                      consumptionLimit:
                        type: object
                        properties:
                          usd:
                            anyOf:
                            - type: number
                              minimum: 0
                            - nullable: true
                              title: 'null'
                            - nullable: true
                              title: 'null'
                            description: USD limit
                            example: 50
                          diem:
                            anyOf:
                            - type: number
                              minimum: 0
                            - nullable: true
                              title: 'null'
                            - nullable: true
                              title: 'null'
                            description: Diem limit
                            example: 10
                          vcu:
                            anyOf:
                            - type: number
                              minimum: 0
                            - nullable: true
                              title: 'null'
                            - nullable: true
                              title: 'null'
                            description: VCU limit (deprecated - use Diem instead)
                            deprecated: true
                            example: 100
                        description: The API Key consumption limits for each epoch.
                        example:
                          usd: 50
                          diem: 10
                      description:
                        type: string
                        description: The API Key description
                        example: Example API Key
                      expiresAt:
                        type: string
                        nullable: true
                        description: The API Key expiration date
                        example: 2023-10-01 12:00:00+00:00
                      id:
                        type: string
                        description: The API Key ID
                        example: e28e82dc-9df2-4b47-b726-d0a222ef2ab5
                    required:
                    - apiKey
                    - apiKeyType
                    - consumptionLimit
                    - expiresAt
                    - id
                    additionalProperties: false
                  success:
                    type: boolean
                required:
                - data
                - success
                additionalProperties: false
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedError'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: An unknown error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
  /api_keys/rate_limits:
    get:
      description: Return details about user balances and rate limits.
      operationId: getApiKeyRateLimits
      summary: /api/v1/api_keys/rate_limits
      tags:
      - API Keys
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      accessPermitted:
                        type: boolean
                        description: Does the API key have access to consume the inference
                          APIs?
                        example: true
                      apiTier:
                        type: object
                        properties:
                          id:
                            type: string
                            description: The ID of the API tier.
                            example: paid
                          isCharged:
                            type: boolean
                            description: Is the API key pay per use (in Diem or USD).
                            example: true
                        required:
                        - id
                        - isCharged
                      balances:
                        type: object
                        properties:
                          USD:
                            type: number
                            description: The USD balance of the key.
                            example: 50.23
                          VCU:
                            type: number
                            description: The VCU balance of the key. VCU is being
                              deprecated in favor of Diem.
                            example: 100.023
                          DIEM:
                            type: number
                            description: The Diem balance of the key.
                            example: 100.023
                      keyExpiration:
                        type: string
                        nullable: true
                        description: The timestamp the API key expires. If null, the
                          key never expires.
                        example: 2025-06-01 00:00:00+00:00
                      nextEpochBegins:
                        type: string
                        description: The timestamp when the next epoch begins. This
                          is relevant for rate limits that reset at the start of each
                          epoch.
                        example: 2025-05-07 00:00:00+00:00
                      rateLimits:
                        type: array
                        items:
                          type: object
                          properties:
                            apiModelId:
                              type: string
                              description: The ID of the API model.
                              example: venice-uncensored
                            rateLimits:
                              type: array
                              items:
                                type: object
                                properties:
                                  amount:
                                    type: number
                                    description: The rate limit for the API model.
                                    example: 100
                                  type:
                                    type: string
                                    description: The time period for the rate limit.
                                      Can be Requests Per Minute (RPM), Requests Per
                                      Day (RPD), or Tokens Per Minute (TPM).
                                    example: RPM
                                required:
                                - amount
                                - type
                          required:
                          - rateLimits
                    required:
                    - accessPermitted
                    - apiTier
                    - balances
                    - keyExpiration
                    - nextEpochBegins
                    - rateLimits
                required:
                - data
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: An unknown error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
  /api_keys/rate_limits/log:
    get:
      description: Returns the last 50 rate limits that the account exceeded.
      operationId: getApiKeyRateLimitLogs
      summary: /api/v1/api_keys/rate_limits/log
      tags:
      - API Keys
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        apiKeyId:
                          type: string
                          description: The ID of the API key that exceeded the limit.
                        modelId:
                          type: string
                          default: venice-uncensored
                          description: The ID of the model that was used when the
                            rate limit was exceeded.
                        rateLimitTier:
                          type: string
                          description: The API tier of the rate limit.
                          example: paid
                        rateLimitType:
                          type: string
                          description: The type of rate limit that was exceeded.
                          example: RPM
                        timestamp:
                          type: string
                          description: The timestamp when the rate limit was exceeded.
                          example: 2023-10-01 12:00:00+00:00
                      required:
                      - apiKeyId
                      - modelId
                      - rateLimitTier
                      - rateLimitType
                      - timestamp
                      additionalProperties: false
                    description: The last 50 rate limit logs for the account.
                  object:
                    type: string
                    enum:
                    - list
                required:
                - data
                - object
                additionalProperties: false
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: An unknown error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
...